refactor FinishTool to builtin tools for agents
