+++ openhands/core/tool/__init__.py
@@ -1,5 +1,6 @@
 """OpenHands runtime package."""
 
+from .builtin_tools import BUILT_IN_TOOLS, FinishTool
 from .tool import ActionBase, ObservationBase, Tool, ToolAnnotations, ToolExecutor
 
 
@@ -9,4 +10,6 @@ __all__ = [
     "ToolExecutor",
     "ActionBase",
     "ObservationBase",
+    "FinishTool",
+    "BUILT_IN_TOOLS",
 ]
