+++ openhands/core/tool/builtin_tools.py
@@ -1,0 +1,55 @@
+"""Implementing essential tools that doesn't interact with the environment.
+
+These are built in and are *required* for the agent to work.
+
+For tools that require interacting with the environment, add them to `openhands/tools`.
+"""
+from pydantic import Field
+
+from openhands.core.tool import ActionBase, ObservationBase, Tool, ToolAnnotations, ToolExecutor
+
+
+class FinishAction(ActionBase):
+    message: str = Field(description="Final message to send to the user.")
+
+class FinishObservation(ObservationBase):
+    message: str = Field(description="Final message sent to the user.")
+
+    @property
+    def agent_observation(self) -> str:
+        return self.message
+
+TOOL_DESCRIPTION = """Signals the completion of the current task or conversation.
+
+Use this tool when:
+- You have successfully completed the user's requested task
+- You cannot proceed further due to technical limitations or missing information
+
+The message should include:
+- A clear summary of actions taken and their results
+- Any next steps for the user
+- Explanation if you're unable to complete the task
+- Any follow-up questions if more information is needed
+"""
+
+class FinishExecutor(ToolExecutor):
+    def __call__(self, action: FinishAction) -> FinishObservation:
+        return FinishObservation(message=action.message)
+
+
+FinishTool = Tool(
+    name="finish",
+    input_schema=FinishAction,
+    output_schema=FinishObservation,
+    description=TOOL_DESCRIPTION,
+    executor=FinishExecutor(),
+    annotations=ToolAnnotations(
+        title="finish",
+        readOnlyHint=True,
+        destructiveHint=False,
+        idempotentHint=True,
+        openWorldHint=False,
+    ),
+)
+
+BUILT_IN_TOOLS = [FinishTool]
